requirements.txt
Django==1.6.5
Pillow==2.4.0
South==0.8.4
django-filebrowser==3.5.6
django-grappelli==2.5.3
django-mptt==0.6.1
django-mptt-admin==0.1.8
djangorestframework==2.3.13
six==1.6.1
wsgiref==0.1.2
